requirements.txt
pypdf
langchain_google_genai
langchain_community
langchain-text-splitters
google-generativeai
chromadb